Fix onboarding v2 route export by moving upload helpers
