+++ app/api/onboarding-v2/sessions/[sessionId]/files/content/route.ts
@@ -1,29 +1,6 @@
 import { NextResponse } from "next/server";
 import { proxyJson, withOnboardingAccess } from "@/features/onboarding-v2/server/apiProxy";
-
-const MAX_BYTES = 10 * 1024 * 1024;
-const SUPPORTED_MIME = new Set(["text/csv", "application/csv", "application/vnd.ms-excel"]);
-
-function parseApproxBase64Bytes(contentBase64: string): number {
-  return Math.floor((contentBase64.length * 3) / 4);
-}
-
-function isCsvLikeExcel(fileName: string): boolean {
-  return fileName.toLowerCase().endsWith(".csv");
-}
-
-function isAllowedUploadType(mimeType: string, fileName: string): { ok: boolean; message?: string } {
-  if (mimeType === "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet") {
-    return { ok: false, message: "xlsx_not_supported_by_agent" };
-  }
-  if (SUPPORTED_MIME.has(mimeType)) {
-    if (mimeType === "application/vnd.ms-excel" && !isCsvLikeExcel(fileName)) {
-      return { ok: false, message: "excel_mime_requires_csv_extension" };
-    }
-    return { ok: true };
-  }
-  return { ok: false, message: "unsupported_file_type" };
-}
+import { isAllowedUploadType, MAX_BYTES, parseApproxBase64Bytes } from "@/features/onboarding-v2/server/fileUpload";
 
 export async function POST(request: Request, context: { params: Promise<{ sessionId: string }> }) {
   const access = await withOnboardingAccess();
@@ -67,4 +44,3 @@ export async function POST(request: Request, context: { params: Promise<{ sessio
   return proxyJson({ method: "POST", path: `/onboarding/sessions/${encodeURIComponent(sessionId)}/files/content`, shopId, body });
 }
 
-export const __testables = { isAllowedUploadType, parseApproxBase64Bytes, MAX_BYTES };
