+++ tests/onboarding-v2/proxy-and-upload.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import { signOnboardingAgentPayload } from "@/features/onboarding-v2/server/signing";
-import { __testables } from "../../app/api/onboarding-v2/sessions/[sessionId]/files/content/route";
+import { isAllowedUploadType, MAX_BYTES, parseApproxBase64Bytes } from "@/features/onboarding-v2/server/fileUpload";
 
 describe("onboarding v2 signing", () => {
   it("signs timestamp.shopId.rawBody format", () => {
@@ -11,11 +11,11 @@ describe("onboarding v2 signing", () => {
 
 describe("upload guardrails", () => {
   it("rejects xlsx as unsupported", () => {
-    expect(__testables.isAllowedUploadType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", "sample.xlsx").ok).toBe(false);
+    expect(isAllowedUploadType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", "sample.xlsx").ok).toBe(false);
   });
 
   it("rejects oversized base64 payload", () => {
-    const bytes = __testables.parseApproxBase64Bytes("A".repeat(__testables.MAX_BYTES * 2));
-    expect(bytes).toBeGreaterThan(__testables.MAX_BYTES);
+    const bytes = parseApproxBase64Bytes("A".repeat(MAX_BYTES * 2));
+    expect(bytes).toBeGreaterThan(MAX_BYTES);
   });
 });
